libs/miroflow/pyproject.toml
[project]
name = "miroflow"
version = "0.1.0"
description = "core functionalities to run llm call and tool call in an agent loop"
readme = "README.md"
authors = [{ name = "MiroMind Team", email = "miromind@shanda.com" }]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.59.0",
    "hydra-core>=1.3.2",
    "json5>=0.12.0",
    "miroflow-contrib>=0.1.0",
    "miroflow-tool>=0.1.0",
    "openai>=1.98.0",
    "pyyaml>=6.0.2",
    # conflict with google-genai
    "rich>=14.1.0",
    "tenacity>=8.2.3,<9.0.0",
    "tiktoken>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/miroflow"]

[tool.uv.sources]
miroflow-tool = { path = "../miroflow-tool", editable = true }
miroflow-contrib = { path = "../miroflow-contrib", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "reuse>=5.0.2",
]
